add cargo mod
