+++ workspace/gh-workflow/src/lib.rs
@@ -1,3 +1,4 @@
+mod cargo;
 pub mod error;
 mod event;
 mod generate;
@@ -5,6 +6,7 @@ mod rust_flag;
 mod toolchain;
 pub(crate) mod workflow;
 
+pub use cargo::*;
 pub use event::*;
 pub use rust_flag::*;
 pub use toolchain::*;
