+++ workspace/gh-workflow/src/workflow.rs
@@ -1,7 +1,6 @@
 #![allow(clippy::needless_update)]
 
 use std::fmt::Display;
-use std::marker::PhantomData;
 
 use derive_setters::Setters;
 use indexmap::IndexMap;
@@ -10,7 +9,7 @@ use serde_json::Value;
 
 use crate::error::Result;
 use crate::generate::Generate;
-use crate::{Event, EventValue, RustFlags, ToolchainStep};
+use crate::{Cargo, Event, EventValue, RustFlags, ToolchainStep};
 
 #[derive(Debug, Default, Setters, Serialize, Deserialize, Clone, PartialEq, Eq)]
 #[serde(rename_all = "kebab-case")]
@@ -91,17 +90,8 @@ impl Workflow {
                     .with_clippy()
                     .with_fmt(),
             )
-            .add_step(
-                Step::cargo(
-                    Cargo::test()
-                        .all_features()
-                        .workspace(),
-                )
-                .name("Cargo Test"),
-            )
-            .add_step(
-                Step::cargo_nightly(Cargo::fmt().check()).name("Cargo Fmt"),
-            )
+            .add_step(Step::cargo(Cargo::test().all_features().workspace()).name("Cargo Test"))
+            .add_step(Step::cargo_nightly(Cargo::fmt().check()).name("Cargo Fmt"))
             .add_step(
                 Step::cargo_nightly(
                     Cargo::clippy()
@@ -184,182 +174,6 @@ pub struct Job {
     pub retry: Option<RetryStrategy>,
     #[serde(skip_serializing_if = "Option::is_none")]
     pub artifacts: Option<Artifacts>,
-}
-
-#[derive(Debug, Serialize, Deserialize, Clone, PartialEq, Eq, strum_macros::Display)]
-#[serde(rename_all = "kebab-case")]
-enum CargoCommand {
-    Add,
-    Bench,
-    Build,
-    Check,
-    Chef,
-    Clean,
-    Clippy,
-    Component,
-    Config,
-    Criterion,
-    Doc,
-    Expand,
-    Fetch,
-    Fix,
-    Flamegraph,
-    Fmt,
-    Generate,
-    GenerateLockfile,
-    GitCheckout,
-    Help,
-    Info,
-    Init,
-    Insta,
-    Install,
-    LocateProject,
-    Login,
-    Logout,
-    Make,
-    Metadata,
-    Miri,
-    New,
-    Owner,
-    Package,
-    ReadManifest,
-    Remove,
-    Report,
-    Run,
-    Rustc,
-    Rustdoc,
-    Search,
-    Test,
-    Tree,
-    Udeps,
-    Uninstall,
-    Update,
-    Vendor,
-    VerifyProject,
-    Version,
-    Watch,
-    Yank,
-}
-#[derive(Debug, PartialEq, Eq)]
-pub struct CargoTest;
-
-#[derive(Debug, PartialEq, Eq)]
-pub struct CargoClippy;
-
-#[derive(Debug, PartialEq, Eq)]
-pub struct CargoFmt;
-
-#[derive(Debug, PartialEq, Eq)]
-pub struct CargoCheck;
-
-#[derive(Debug, PartialEq, Eq)]
-pub struct CargoClean;
-
-#[derive(Debug, PartialEq, Eq)]
-pub struct Cargo<Ty> {
-    command: CargoCommand,
-    args: Vec<String>,
-
-    marker: PhantomData<Ty>,
-}
-
-impl<Ty> Cargo<Ty> {
-    // TODO: add function to handle command
-    // specific sub command
-    pub fn add_arg<S: ToString>(mut self, arg: S) -> Self {
-        self.args.push(arg.to_string());
-
-        self
-    }
-}
-
-impl Cargo<CargoClean> {
-    pub fn clean() -> Cargo<CargoClean> {
-        Cargo {
-            command: CargoCommand::Clean,
-            args: vec![],
-            marker: Default::default(),
-        }
-    }
-}
-
-impl Cargo<CargoCheck> {
-    pub fn check() -> Cargo<CargoCheck> {
-        Cargo {
-            command: CargoCommand::Check,
-            args: vec![],
-            marker: Default::default(),
-        }
-    }
-}
-
-impl Cargo<CargoTest> {
-    pub fn test() -> Cargo<CargoTest> {
-        Cargo {
-            command: CargoCommand::Test,
-            args: vec![],
-            marker: Default::default(),
-        }
-    }
-    pub fn all_features(mut self) -> Self {
-        if !self.args.iter().any(|arg| arg == "--all-features") {
-            self.args.push("--all-features".to_string());
-        }
-
-        self
-    }
-    pub fn workspace(mut self) -> Self {
-        if !self.args.iter().any(|arg| arg == "--workspace") {
-            self.args.push("--workspace".to_string());
-        }
-        self
-    }
-}
-
-
-impl Cargo<CargoClippy> {
-    pub fn clippy() -> Cargo<CargoClippy> {
-        Cargo {
-            command: CargoCommand::Clippy,
-            args: vec![],
-            marker: Default::default(),
-        }
-    }
-    pub fn all_features(mut self) -> Self {
-        if !self.args.iter().any(|arg| arg == "--all-features") {
-            self.args.push("--all-features".to_string());
-        }
-        self
-    }
-    pub fn workspace(mut self) -> Self {
-        if !self.args.iter().any(|arg| arg == "--workspace") {
-            self.args.push("--workspace".to_string());
-        }
-        self
-    }
-}
-
-impl Cargo<CargoFmt> {
-    pub fn fmt() -> Cargo<CargoFmt> {
-        Cargo {
-            command: CargoCommand::Fmt,
-            args: vec![],
-            marker: Default::default(),
-        }
-    }
-    pub fn all(mut self) -> Self {
-        if !self.args.iter().any(|arg| arg == "--all") {
-            self.args.push("--all".to_string());
-        }
-
-        self
-    }
-    pub fn check(mut self) -> Self {
-        if !self.args.iter().any(|arg| arg == "--check") {
-            self.args.push("--check".to_string());
-        }
-        self
-    }
 }
 
 impl Job {
